feat: add opengov scenario
