+++ src/data/opengov_data.json
@@ -1,0 +1,79 @@
+{
+  "chains": [
+    {
+      "name": "Polkadot",
+      "chainId": "91b171bb158e2d3848fa23a9f1c25182fb8e20313b2c1eb49219da7a70ce90c3",
+      "referendumIds": [
+        1140,
+        1146,
+        1147,
+        1148,
+        1149,
+        1182,
+        1158,
+        1198,
+        1162,
+        1187,
+        1190,
+        1191,
+        1192,
+        1208,
+        1206,
+        1163,
+        1196,
+        1169,
+        1199,
+        1157,
+        1171,
+        1205,
+        1173,
+        1175,
+        1176,
+        1207,
+        1160,
+        1166,
+        1177,
+        1209,
+        1180,
+        1183,
+        1184,
+        1185,
+        1186,
+        1189,
+        1193,
+        1197,
+        1200,
+        1201,
+        1203,
+        1204,
+        1202,
+        1195
+      ]
+    },
+    {
+      "name": "Kusama",
+      "chainId": "b0a8d493285c2df73290dfb7e61f870f17b41801197a149ca93654499ea3dafe",
+      "referendumIds": [
+        462,
+        460,
+        461,
+        458
+      ]
+    }
+  ],
+  "languageIsoCode": [
+    "en",
+    "zh",
+    "it",
+    "pt",
+    "ru",
+    "es",
+    "tr",
+    "fr",
+    "in",
+    "pl",
+    "ja",
+    "vi",
+    "ko"
+  ]
+}
