+++ src/tasks/__init__.py
@@ -1,0 +1,3 @@
+from .governance_task import GovernanceTasks
+from .history_task import HistoryTasks
+from .multistaking_task import MultiStakingTasks
